+++ packages/test/harness/src/host/common/host/sendDevToolsCommand.js
@@ -1,0 +1,5 @@
+module.exports = webDriver => {
+  return function sendDevToolsCommand(command, options) {
+    return webDriver.sendDevToolsCommand(command, options);
+  };
+};